Add WPILib as a packaged jar to the project
This commit adds WPILib and updates some build files
so that they point to the local copy rather than the one installed
at user.home. Note that the local version of WPILib is not the same
as the one installed at user.home as it has modified buildfiles and
removes the javadocs (for space).

Note that this removes a line from the .gitignore, specificlly,
it removes the one which ignores jar files.