Make the build target `download-wpilib` download from a git repository URL.
Why? Well, wpilib requires ALL user library jars to be placed in wpilib/user/java/lib,
which is a folder that this repo does not "own". Additionally, there is no easy
or elegant way to move jars into this directory. The simplest solution I can
think of is copying jars into the file from this repo, but I really don't want
to include too many jar files in this repo. Addionally, the site that was
previously downloading WPILib did not include the cpp or tools folders (which
might be nessecarry in order to deploy, I'm not sure).

Instead, I decided to download the wpilib folder directly from a repo that we
"own" (currently it's set to a repo I just made, but we can switch it over to
the FRC1076 organization). As a bonus, this means that updating WPILib is easy,
you update WPILib on a machine through eclipse, and then you push these changes
to GitHub. Then, anyone can pull these updates without having to go through the
same troubles. It also means that dependencies are easier to share between
projects (and only one person needs to ever download these dependencies).