Change CANTalon import from WPILib to the CTRE library
In the new 2017 WPILib, all Talon motor related classes have been moved to an external userlibrary called "com.ctre". This commit changes the only instance of a CANTalon to be imported from the com.ctre library. The classpath has also been updated accordingly. In addition, the JSON library was also added to the WPILib user folder (and the classpath updated to reflect that. The old entry is still there for the time being)